# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SenAI CRM Agentic Intelligence Platform API
  version: 1.0.0
  description: |
    Production-grade AI-powered CRM with autonomous email triage, RAG pipeline,
    ReAct agent, and real-time analytics.

    ## Authentication
    All endpoints require no auth in dev mode. Add `Authorization: Bearer <token>` header in production.

    ## Error Format
    All errors follow a consistent envelope:
    ```json
    { "error_code": "NOT_FOUND", "message": "Email not found", "details": {} }
    ```
  contact:
    email: support@senai.io
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.senai.io
    description: Production server
...
paths:
...
  /api/v1/status/{job_id}:
    get:
      tags: [ingestion]
      summary: Check email processing status
      operationId: getJobStatus
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: integer
          example: 42
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatusResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─────────────── THREADS ───────────────
  /api/v1/threads:
    get:
      tags: [threads]
      summary: List all threads
      operationId: listThreads
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, escalated, resolved, legal_flagged, ignored]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of threads
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  threads:
                    type: array
                    items:
                      $ref: '#/components/schemas/ThreadSummary'
  /api/v1/threads/{thread_id}:
    get:
      tags: [threads]
      summary: Get full thread with all emails and agent reasoning
      operationId: getThread
      parameters:
        - name: thread_id
          in: path
          required: true
          schema:
            type: string
          example: thread_bob_outage
      responses:
        '200':
          description: Thread with full email history and agent steps
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadDetail'
        '404':
          description: Thread not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/emails/{email_id}:
    get:
      tags: [threads]
      summary: Get a single email with classification details
      operationId: getEmail
      parameters:
        - name: email_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Email detail with agent decision and RAG context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailDetail'
        '404':
          description: Email not found
  /api/v1/emails/{email_id}/draft:
    patch:
      tags: [threads]
      summary: Edit the AI-generated draft reply
      operationId: editDraft
      parameters:
        - name: email_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [draft_reply]
              properties:
                draft_reply:
                  type: string
                  example: "Dear Bob, we sincerely apologize for the SLA breach..."
      responses:
        '200':
          description: Draft updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  draft_reply:
                    type: string
  /api/v1/emails/{email_id}/approve:
    post:
      tags: [threads]
      summary: Approve and send the draft reply
      description: Triggers the human-in-the-loop approval. Marks email as replied and logs to audit trail.
      operationId: approveDraft
      parameters:
        - name: email_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                approved:
                  type: boolean
                  default: true
      responses:
        '200':
          description: Draft approved and sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  email_id:
                    type: string
                  approved:
                    type: boolean

  # ─────────────── CONTACTS ───────────────
  /api/v1/contacts:
    get:
      tags: [contacts]
      summary: List all contacts
      operationId: listContacts
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [VIP, Active, Blocked, Churned]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of contacts
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  contacts:
                    type: array
                    items:
                      $ref: '#/components/schemas/ContactSummary'
  /api/v1/contacts/{email}:
    get:
      tags: [contacts]
      summary: Get contact profile with churn risk and thread history
      operationId: getContact
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
          example: karen.w@retail-co.com
      responses:
        '200':
          description: Contact profile with full CRM data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactDetail'
        '404':
          description: Contact not found
  /api/v1/contacts/{email}/status:
    patch:
      tags: [contacts]
      summary: Update contact status (VIP, Blocked, etc.)
      operationId: updateContactStatus
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [VIP, Active, Blocked, Churned]
      responses:
        '200':
          description: Status updated
        '404':
          description: Contact not found

  # ─────────────── AGENT ───────────────
  /api/v1/agent/triage/manual:
    post:
      tags: [agent]
      summary: Manually trigger agent triage on a new email
      description: |
        Runs the full ReAct agent loop on the provided email. Useful for testing
        specific scenarios without ingesting through the pipeline.
      operationId: manualTriage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualTriageRequest'
            examples:
              bob_escalation:
                summary: Bob SLA Breach (msg_060)
                value:
                  email_id: msg_060
                  subject: "Escalation: SLA Breach + Legal Review"
                  body: "Our legal team is now involved. Please expect formal correspondence. We are putting the renewal on hold."
                  sender_email: bob.jones@enterprise.net
                  contact_name: Bob Jones
                  dry_run: true
      responses:
        '200':
          description: Agent triage result with full reasoning trace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentTriageResult'
...
  /api/v1/rag/search:
    get:
      tags: [rag]
      summary: Search the knowledge base (debug endpoint)
      description: |
        Query the ChromaDB vector database for relevant policy chunks.
        Returns top-N chunks with similarity scores.
      operationId: ragSearch
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          example: "refund policy 14 days non-profit discount"
        - name: n
          in: query
          schema:
            type: integer
            default: 3
      responses:
        '200':
          description: Retrieved chunks with similarity scores
          content:
            application/json:
              schema:
                type: object
                properties:
                  query:
                    type: string
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/RAGChunk'
                  total_kb_docs:
                    type: integer
...
components:
  schemas:
...
    JobStatusResponse:
      type: object
      properties:
        job_id:
          type: integer
        email_id:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed]
        classification:
          type: string
          nullable: true
        agent_decision:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    ThreadSummary:
      type: object
      properties:
        id:
          type: string
        external_id:
          type: string
        subject:
          type: string
        status:
          type: string
        priority_score:
          type: number
        email_count:
          type: integer
        last_decision:
          type: string
          nullable: true
        updated_at:
          type: string
          format: date-time
    ThreadDetail:
      type: object
      properties:
        id:
          type: string
        external_id:
          type: string
        subject:
          type: string
        status:
          type: string
        priority_score:
          type: number
        sentiment_trend:
          type: object
          nullable: true
        contact:
          $ref: '#/components/schemas/ContactSummary'
        emails:
          type: array
          items:
            $ref: '#/components/schemas/EmailDetail'
    EmailDetail:
      type: object
      properties:
        id:
          type: string
        external_id:
          type: string
          nullable: true
        subject:
          type: string
          nullable: true
        body:
          type: string
          nullable: true
        sender_email:
          type: string
          nullable: true
        received_at:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
        classification:
          type: string
          nullable: true
        sentiment:
          type: string
          nullable: true
        confidence:
          type: number
          nullable: true
        agent_decision:
          type: string
          nullable: true
          enum: [Auto-Reply, Escalate, Legal-Flag, Human-Review, Ignore]
        agent_steps:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/AgentStep'
        draft_reply:
          type: string
          nullable: true
        draft_approved:
          type: boolean
        ingested_at:
          type: string
          format: date-time
    AgentStep:
      type: object
      properties:
        step:
          type: integer
        type:
          type: string
          enum: [classify, decide, react, final_decision]
        tool:
          type: string
          nullable: true
        thought:
          type: string
        result:
          type: object
    ContactSummary:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
          nullable: true
        company:
          type: string
          nullable: true
        tier:
          type: string
        ltv:
          type: number
        churn_risk:
          type: number
          nullable: true
    ContactDetail:
      allOf:
        - $ref: '#/components/schemas/ContactSummary'
        - type: object
          properties:
            status:
              type: string
            tags:
              type: array
              items:
                type: string
            notes:
              type: string
              nullable: true
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            threads:
              type: array
              items:
                $ref: '#/components/schemas/ThreadSummary'
    ManualTriageRequest:
      type: object
      required: [email_id, subject, body, sender_email]
      properties:
        email_id:
          type: string
          example: msg_060
        subject:
          type: string
          example: "Escalation: SLA Breach + Legal Review"
        body:
          type: string
        sender_email:
          type: string
          format: email
        contact_name:
          type: string
          nullable: true
        contact_company:
          type: string
          nullable: true
        dry_run:
          type: boolean
          default: false
          description: "If true, agent shows plan but does not execute tool side-effects"
    AgentTriageResult:
      type: object
      properties:
        email_id:
          type: string
        decision:
          type: string
          enum: [Auto-Reply, Escalate, Legal-Flag, Human-Review, Ignore]
        classification:
          type: string
        sentiment:
          type: string
        confidence:
          type: number
        draft_reply:
          type: string
          nullable: true
        steps:
          type: array
          items:
            $ref: '#/components/schemas/AgentStep'
        processed_at:
          type: string
          format: date-time
...
    RAGChunk:
      type: object
      properties:
        id:
          type: string
        document:
          type: string
          description: The retrieved policy text chunk
        metadata:
          type: object
          properties:
            source:
              type: string
              example: refund_policy
            category:
              type: string
              example: refund
            chunk_index:
              type: integer
        score:
          type: number
          description: Cosine similarity score (0.0 - 1.0)
          example: 0.847
...
    ErrorResponse:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          type: string
          example: NOT_FOUND
          enum:
            - NOT_FOUND
            - DUPLICATE_ID
            - VALIDATION_ERROR
            - INTERNAL_ERROR
            - BLOCKLISTED_SENDER
            - KB_UNAVAILABLE
        message:
          type: string
          example: "Email not found"
        details:
          type: object
          additionalProperties: true